+++ src/sections/userpreset-browser.ts
@@ -57,7 +57,7 @@ export class UserPresetBrowser extends FavBrowserBase {
 
     // render html.
     return html`
-      <div class="media-browser-section">
+      <div class="media-browser-section" style=${this.styleMediaBrowser()}>
         ${title ? html`<div class="media-browser-section-title">${title}</div>` : html``}
         ${subtitle ? html`<div class="media-browser-section-subtitle">${subtitle}</div>` : html``}
         <div class="media-browser-controls">